Refactor hotel-related components and update dependencies: introduce new UI components for hotel search and selection, modify existing data classes and repository methods for improved API integration, and update Gradle configurations.
solucio d'errors que impedeixen la compilacio correcta de l'app

+++ app/src/main/java/com/example/waygo/data/remote/model/Room.kt
@@ -3,6 +3,6 @@ package com.example.waygo.data.remote.model
 data class Room(
     val id: String,
     val roomType: String,
-    val price: Double,
+    val price: Float,
     val images: List<String>
 )
